Dev: Add abstract TestAuthDirectory class
