+++ addOns/dev/src/main/java/org/zaproxy/addon/dev/auth/passwordNewPage/PasswordNewPageDir.java
@@ -19,45 +19,19 @@
  */
 package org.zaproxy.addon.dev.auth.passwordNewPage;
 
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Map;
-import org.apache.commons.lang.RandomStringUtils;
-import org.zaproxy.addon.dev.TestDirectory;
+import org.zaproxy.addon.dev.TestAuthDirectory;
 import org.zaproxy.addon.dev.TestProxyServer;
 
 /**
  * A login page which uses one JSON request to login endpoint. The token is returned in a standard
  * field.
  */
-public class PasswordNewPageDir extends TestDirectory {
-
-    // These are test credentials, so hardcoding them is fine ;)
-    private static final String[][] USERS = {{"test@test.com", "password123"}};
-
-    private Map<String, String> sessions = new HashMap<>();
+public class PasswordNewPageDir extends TestAuthDirectory {
 
     public PasswordNewPageDir(TestProxyServer server, String name) {
         super(server, name);
         this.addPage(new PasswordNewPageLoginPage(server));
         this.addPage(new PasswordNewPageNextPage(server));
         this.addPage(new PasswordNewPageVerificationPage(server));
-    }
-
-    public boolean isValid(String username, String password) {
-        return Arrays.stream(USERS)
-                .filter(c -> (c[0].equals(username) && c[1].equals(password)))
-                .findAny()
-                .isPresent();
-    }
-
-    public String getToken(String username) {
-        String token = RandomStringUtils.randomAlphanumeric(32);
-        sessions.put(token, username);
-        return token;
-    }
-
-    public String getUser(String token) {
-        return sessions.get(token);
     }
 }
